# Use Object#send in order to increase performance